requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
elevenlabs
notebook
requests 